client GL: fix ARB_fragment_program state
We always use the fragment shader now (since RGB rendering is done without GL), so the default state should be ENABLED. This fixes cases where apitrace showed me that we were rendering a YUV frame with the shader disabled. This may also have stemmed from an incorrect order of glEnable/glBindProgram calls, which this patch changes as well.

git-svn-id: https://www.xpra.org/svn/Xpra/trunk/src@2948 3bb7dfac-3a0b-4e04-842a-767bc560f471